Version 1.2.3
Added global variable for pool border wall distance. Adjusted all
functions for that macro.
Fixed the redo function for the pool maize to include the diameter
parameter.
Added the number of regions/quadrants for NOR/WM in heatmap/linetrack.
Fixed the ambiguity in OF about displacement in center and outer+border
regions.
Added a global variable for minimum value to consider a mouse to be
stopped.
Checked the accuracy of explorations in EM.
Removed references to doublets in TY.
Corrected the freezed parameters in FC.
Added drawing of border in SM.